Add option to omit scope in TTS messages
When a user sends a message to the log, the TTS engine
will play "Channel <username> <message>". When omit scope
box is selected, the phrase "Channel" is omitted from the TTS
engine which allows for the users message to quickly be spoken
for all clients. There is also an option not say the username.

v2: Refactor to avoid string parsing
v3: Rename variables and setting string
v4: Cleaned up Log.ui